requirements.txt
torch
torchvision
matplotlib
numpy
Pillow
typed-argument-parser
timm==0.5.4
tensorboardx
pytz
tensorboard
setuptools==59.5.0
pandas
wandb
opencv-python
scikit-image